change behaviour to die if reaching at the same time

+++ Codewars/8kyu/Holiday4SharkPontoon/Holiday4SharkPontoon.Logic/Kata.cs
@@ -3,7 +3,7 @@
 public class Kata
 {
     public static string Shark(int pontoonDistance, int sharkDistance, int yourSpeed, int sharkSpeed, bool dolphin) =>
-        pontoonDistance / (float)yourSpeed < sharkDistance * (dolphin? 2.0 : 1.0) / sharkSpeed
+        pontoonDistance / (double)yourSpeed < sharkDistance * (dolphin? 2.0 : 1.0) / sharkSpeed
             ? "Alive!"
             : "Shark Bait!";
 }
